fixes
- move some code
- make event audit entirely redux/store driven
- remove animations
