+++ dispatch/static/manager/src/js/actions/EventsActions.js
@@ -1,4 +1,5 @@
 import { push } from 'react-router-redux'
+import { normalize } from 'normalizr'
 
 import * as types from '../constants/ActionTypes'
 import { eventSchema } from '../constants/Schemas'
@@ -18,6 +19,26 @@ class EventsActions extends ResourceActions {
 
     return dispatch => {
       dispatch(push({ pathname: '/events/', query: queryObj }))
+    }
+  }
+
+  approve(token, id, queryAfter) {
+    const form = new FormData()
+    form.append('is_submission', false)
+
+    return dispatch => {
+      dispatch(() => dispatch({
+        type: types.EVENTS.SAVE,
+        payload: DispatchAPI.events.save(token, id, form)
+          .then(json => {
+            if (queryAfter) {
+              dispatch(this.list(token, queryAfter))
+            }
+            return {
+              data: normalize(json, eventSchema)
+            }
+          })
+      }))
     }
   }
