+++ dispatch/static/manager/src/js/pages/Events/index.js
@@ -1,9 +1,11 @@
 import EventIndexPage from './EventIndexPage'
 import EventPage from './EventPage'
 import NewEventPage from './NewEventPage'
+import Audit from './Audit'
 
 export {
   EventIndexPage as Index,
   EventPage as Event,
-  NewEventPage as NewEvent
+  NewEventPage as NewEvent,
+  Audit
 }
