+++ dispatch/static/manager/src/js/util/helpers.js
@@ -1,4 +1,5 @@
 import moment from 'moment'
+import R from 'ramda'
 
 export function humanizeDatetime(timestamp) {
   return moment(timestamp).format('lll') // April 18, 2017 12:50 PM
@@ -9,4 +10,15 @@ export function dateObjToAPIString(date) {
     return date.toISOString()
   }
   return date
+}
+
+export function getPath(location, page) {
+  let query = R.clone(location.query)
+
+  query.page = page
+
+  return {
+    pathname: location.pathname,
+    query: query
+  }
 }
